+++ lib/graphql/types/relay.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
+require "graphql/types/relay/default_relay"
 require "graphql/types/relay/connection_behaviors"
-require "graphql/types/relay/base_field"
+require "graphql/types/relay/node_behaviors"
 require "graphql/types/relay/base_object"
-require "graphql/types/relay/base_interface"
 require "graphql/types/relay/page_info"
 require "graphql/types/relay/base_connection"
 require "graphql/types/relay/base_edge"
