+++ lib/graphql/types/relay/base_object.rb
@@ -4,7 +4,6 @@ module GraphQL
   module Types
     module Relay
       class BaseObject < GraphQL::Schema::Object
-        field_class(Types::Relay::BaseField)
         class << self
           def default_relay(new_value)
             @default_relay = new_value
