+++ lib/graphql/types/relay/node.rb
@@ -7,10 +7,8 @@ module GraphQL
       # or you can take it as inspiration for your own implementation
       # of the `Node` interface.
       module Node
-        include Types::Relay::BaseInterface
-        default_relay(true)
-        description "An object with an ID."
-        field(:id, ID, null: false, description: "ID of the object.")
+        include GraphQL::Schema::Interface
+        include Types::Relay::NodeBehaviors
       end
     end
   end
